uf-10211: Move open-e model package.
Also consolidate "Berakningar(Intern/Extern)".
