+++ src/main/java/se/sundsvall/billingdatacollector/integration/opene/mapper/kundfakturaformular/ExternalConfig.java
@@ -5,16 +5,16 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Scope;
 
-import se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.external.AktivitetskontoExtern;
-import se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.external.AnsvarExtern;
-import se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.external.BerakningExtern;
-import se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.external.BarakningarExtern;
-import se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.external.MomssatsExtern;
-import se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.external.ObjektkontoExtern;
-import se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.external.ProjektkontoExtern;
-import se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.external.SummeringExtern;
-import se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.external.UnderkontoExtern;
-import se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.external.VerksamhetExtern;
+import se.sundsvall.billingdatacollector.integration.opene.model.external.AktivitetskontoExtern;
+import se.sundsvall.billingdatacollector.integration.opene.model.external.AnsvarExtern;
+import se.sundsvall.billingdatacollector.integration.opene.model.external.BerakningExtern;
+import se.sundsvall.billingdatacollector.integration.opene.model.external.BarakningarExtern;
+import se.sundsvall.billingdatacollector.integration.opene.model.external.MomssatsExtern;
+import se.sundsvall.billingdatacollector.integration.opene.model.external.ObjektkontoExtern;
+import se.sundsvall.billingdatacollector.integration.opene.model.external.ProjektkontoExtern;
+import se.sundsvall.billingdatacollector.integration.opene.model.external.SummeringExtern;
+import se.sundsvall.billingdatacollector.integration.opene.model.external.UnderkontoExtern;
+import se.sundsvall.billingdatacollector.integration.opene.model.external.VerksamhetExtern;
 
 /**
  * Configuration class for external OpenE objects
