+++ src/main/java/se/sundsvall/billingdatacollector/integration/opene/model/external/BarakningarExtern.java
@@ -1,4 +1,4 @@
-package se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.external;
+package se.sundsvall.billingdatacollector.integration.opene.model.external;
 
 import lombok.Builder;
 import lombok.Getter;
