+++ src/main/java/se/sundsvall/billingdatacollector/integration/opene/model/external/BerakningExtern.java
@@ -1,4 +1,4 @@
-package se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.external;
+package se.sundsvall.billingdatacollector.integration.opene.model.external;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -9,8 +9,10 @@ import lombok.ToString;
 @Getter
 @Builder(setterPrefix = "with")
 @ToString
-public class VerksamhetExtern {
+public class BerakningExtern {
 	private String queryID;
 	private String name;
-	private String value;
+	private String fakturatextExtern;
+	private String antalExtern;
+	private String aPrisExtern;
 }
