+++ src/main/java/se/sundsvall/billingdatacollector/integration/opene/model/external/MomssatsExtern.java
@@ -1,4 +1,4 @@
-package se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.internal;
+package se.sundsvall.billingdatacollector.integration.opene.model.external;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -9,7 +9,7 @@ import lombok.ToString;
 @Getter
 @Builder(setterPrefix = "with")
 @ToString
-public class ProjektkontoIntern {
+public class MomssatsExtern {
 	private String queryID;
 	private String name;
 	private String value;
