+++ src/main/java/se/sundsvall/billingdatacollector/integration/opene/model/external/SummeringExtern.java
@@ -1,4 +1,4 @@
-package se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.internal;
+package se.sundsvall.billingdatacollector.integration.opene.model.external;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -9,8 +9,8 @@ import lombok.ToString;
 @Getter
 @Builder(setterPrefix = "with")
 @ToString
-public class SummeringIntern {
+public class SummeringExtern {
 	private String queryID;
 	private String name;
-	private String totSummeringIntern;
+	private String totSummeringExtern;
 }
