+++ src/main/java/se/sundsvall/billingdatacollector/integration/opene/model/external/UnderkontoExtern.java
@@ -1,4 +1,4 @@
-package se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.external;
+package se.sundsvall.billingdatacollector.integration.opene.model.external;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -9,7 +9,7 @@ import lombok.ToString;
 @Getter
 @Builder(setterPrefix = "with")
 @ToString
-public class AktivitetskontoExtern {
+public class UnderkontoExtern {
 	private String queryID;
 	private String name;
 	private String value;
