+++ src/main/java/se/sundsvall/billingdatacollector/integration/opene/model/internal/AktivitetskontoIntern.java
@@ -1,4 +1,4 @@
-package se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.internal;
+package se.sundsvall.billingdatacollector.integration.opene.model.internal;
 
 import lombok.Builder;
 import lombok.Getter;
