+++ src/main/java/se/sundsvall/billingdatacollector/integration/opene/model/internal/AnsvarIntern.java
@@ -1,4 +1,4 @@
-package se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.external;
+package se.sundsvall.billingdatacollector.integration.opene.model.internal;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -9,7 +9,7 @@ import lombok.ToString;
 @Getter
 @Builder(setterPrefix = "with")
 @ToString
-public class MomssatsExtern {
+public class AnsvarIntern {
 	private String queryID;
 	private String name;
 	private String value;
