+++ src/main/java/se/sundsvall/billingdatacollector/integration/opene/model/internal/ProjektkontoIntern.java
@@ -1,4 +1,4 @@
-package se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.internal;
+package se.sundsvall.billingdatacollector.integration.opene.model.internal;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -9,7 +9,7 @@ import lombok.ToString;
 @Getter
 @Builder(setterPrefix = "with")
 @ToString
-public class UnderkontoIntern {
+public class ProjektkontoIntern {
 	private String queryID;
 	private String name;
 	private String value;
