+++ src/main/java/se/sundsvall/billingdatacollector/integration/opene/model/internal/SummeringIntern.java
@@ -1,4 +1,4 @@
-package se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.external;
+package se.sundsvall.billingdatacollector.integration.opene.model.internal;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -9,8 +9,8 @@ import lombok.ToString;
 @Getter
 @Builder(setterPrefix = "with")
 @ToString
-public class ProjektkontoExtern {
+public class SummeringIntern {
 	private String queryID;
 	private String name;
-	private String value;
+	private String totSummeringIntern;
 }
