+++ src/main/java/se/sundsvall/billingdatacollector/integration/opene/model/internal/UnderkontoIntern.java
@@ -1,4 +1,4 @@
-package se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.internal;
+package se.sundsvall.billingdatacollector.integration.opene.model.internal;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -9,7 +9,7 @@ import lombok.ToString;
 @Getter
 @Builder(setterPrefix = "with")
 @ToString
-public class AnsvarIntern {
+public class UnderkontoIntern {
 	private String queryID;
 	private String name;
 	private String value;
