+++ src/main/java/se/sundsvall/billingdatacollector/integration/opene/model/internal/VerksamhetIntern.java
@@ -1,0 +1,16 @@
+package se.sundsvall.billingdatacollector.integration.opene.model.internal;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+@Setter
+@Getter
+@Builder(setterPrefix = "with")
+@ToString
+public class VerksamhetIntern {
+	private String queryID;
+	private String name;
+	private String value;
+}
