+++ src/main/java/se/sundsvall/billingdatacollector/integration/opene/util/ListUtil.java
@@ -18,7 +18,7 @@ import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 import org.xml.sax.SAXException;
 
-import se.sundsvall.billingdatacollector.integration.opene.mapper.kundfakturaformular.model.OpeneCollections;
+import se.sundsvall.billingdatacollector.integration.opene.model.OpeneCollections;
 
 @Component
 public class ListUtil {
